fedimint_rocksdb::rocks_db_open_blocking_builder

Trait State

pub trait State: Sized {
    type RelaxedConsistency;
}

Builder’s type state specifies if members are set or not (unset).

You can use the associated types of this trait to control the state of individual members with the IsSet and IsUnset traits. You can change the state of the members with the Set* structs available in this module.

Required Associated Types

type RelaxedConsistency

Type state of the member relaxed_consistency.

It can implement either IsSet or IsUnset

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.